% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam.probe.R
\name{gam.probe}
\alias{gam.probe}
\title{Probe an interaction with GAM}
\usage{
gam.probe(
  x,
  z,
  y,
  k,
  zs = NULL,
  spotlights = NULL,
  xlab = "moderator",
  col1 = "red4",
  col2 = "dodgerblue",
  coldy = "purple",
  ylab1 = "Dependent Variable",
  ylab2 = "Marginal Effect",
  main1 = "Simple Slopes",
  main2 = "Floodlight",
  ...
)
}
\arguments{
\item{x}{the predictor of interest (in an experiment, the discrete randomly
assigned manipulation).}

\item{z}{the moderator}

\item{y}{the dependent variable.}

\item{col1}{color used when x takes its lowest value}

\item{col2}{color used when x takes its highest value}
}
\description{
The interaction is probed as proposed in Simonsohn (2024), estimating a GAM model
and computing both simple slopes and floodlight/Johnson-Neyman procedure.
}
