% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interprobe.R
\name{interprobe}
\alias{interprobe}
\title{Probe interactions computing Simple Slopes and Floodlight (Johnson-Neyman)}
\usage{
interprobe(
  model = NULL,
  x,
  z,
  y,
  data = NULL,
  k = NULL,
  zs = NULL,
  spotlights = NULL,
  plot = TRUE,
  histogram = TRUE,
  xlab = "moderator",
  col1 = "red4",
  col2 = "dodgerblue",
  col3 = "purple",
  ylab1 = "Dependent Variable",
  ylab2 = "Marginal Effect",
  main1 = "GAM Simple Slopes",
  main2 = "GAM Floodlight",
  ...
)
}
\arguments{
\item{x}{the predictor of interest (in an experiment, the discrete randomly
assigned manipulation).}

\item{z}{the moderator}

\item{y}{the dependent variable.}

\item{col1}{color used when x takes its lowest value}

\item{col2}{color used when x takes its highest value}
}
\description{
The interaction is probed, by default, as proposed in Simonsohn (2024), estimating a GAM model
and computing both simple slopes and floodlight/Johnson-Neyman procedure.
}
