% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interprobe_dev.R
\name{interprobe_dev}
\alias{interprobe_dev}
\title{Probe interactions computing Simple Slopes and Floodlight (Johnson-Neyman)}
\usage{
interprobe_dev(
  x = NULL,
  z = NULL,
  y = NULL,
  data = NULL,
  model = NULL,
  k = NULL,
  zs = NULL,
  spotlights = NULL,
  draw = TRUE,
  histogram = TRUE,
  xlab = "moderator",
  col1 = "red4",
  col2 = "dodgerblue",
  col3 = "purple",
  ylab1 = "Dependent Variable",
  ylab2 = "Marginal Effect",
  main1 = "GAM Simple Slopes",
  main2 = "GAM Floodlight",
  ...
)
}
\arguments{
\item{x}{the predictor of interest (in an experiment, the discrete randomly
assigned manipulation). Can be the name of a variable (e.g., x='treatment') or data (e.g., x=c(1,3,1,4)}

\item{z}{the moderator. Can be the name of variable, or a vector with data}

\item{y}{the dependent variable. Can be the name of variable, or a vector with data}

\item{draw}{TRUE/FALSE for whether to make a plot}

\item{col1}{color used when x takes its lowest value}

\item{col2}{color used when x takes its highest value}
}
\description{
The interaction is probed, by default, as proposed in Simonsohn (2024), estimating a GAM model
and computing both simple slopes and floodlight/Johnson-Neyman procedure.
}
