% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.split.R
\name{lm.split}
\alias{lm.split}
\title{Interrupted regression to assess robustness of GAM-probing of an interaction}
\usage{
lm.split(x, z, y, zc, data, quiet = FALSE)
}
\arguments{
\item{x}{dichotomous variable indicating randomly assigned condition.
Can be the name of a variable (e.g., x='treatment') when
providing a data argument, or a vector with the x values.}

\item{z}{the moderator. Same options as for \code{x}}

\item{y}{the dependent variable. Same options as for \code{x}.}

\item{zc}{cutoff point for the moderator where the two  regression lines will be interrupted.}

\item{data}{an optional data frame containing the focal predictor, moderator and
dependent variable. If \code{data} is specified, \code{x}, \code{z} and \code{y} must be names of
variables in that dataframe.}

\item{quiet}{set to TRUE to reduce the output shown on console as lm.split() executes.}
}
\description{
After a pattern of interest is documented with GAM probing, one conducts
an interrupted regression estimating the effect of the key predictor, x, and
the moderator, z, on y, allowing for different slopes and interactions across two
range of values of the moderator. This approach is a valid robustness check only
in situations where we can expect r(x,z)=0, so primarily when either x or z
are randomly assigned.
}
